feat: Add Configuration Explorer page
Introduces a new page for visualizing and exploring application configurations. This includes adding the new route and page component, integrating Chart.js for data visualization, and updating the sidebar navigation to include a link to the new page. Dependencies for Font Awesome and Chart.js are also added to the index.html.